requirements.txt
setuptools>=78.0.2
wheel>=0.45.1
python-telegram-bot==20.7
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pytz==2024.1
jdatetime==4.1.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.18.0
kaleido==0.2.1 